fixed should been uint not byte

+++ Assets/Mirage/Runtime/INetworkIdentityGenerator.cs
@@ -5,6 +5,6 @@ namespace Mirage
         /// <summary>
         ///     Generate your own specific net id.
         /// </summary>
-        byte GenerateNetId();
+        uint GenerateNetId();
     }
 }
